Initial implementation of proxy support.
- support a proxy method that allows Java to call a global JavaScript
  object.
- still a work in progress: no argument support, only strings for return
  values.
- misc: split up the tests into separate fixtures by functional group.
